vistara_backend/requirements.txt
Django==6.0
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
redis
celery
django-celery-results
Pillow
requests
django-cors-headers
